+++ models/Group.go
@@ -1,7 +1,7 @@
 package models
 
 type Group struct {
-    ID         int    `json:"id" xorm:"unique"`
+    ID         int64    `json:"id" xorm:"unique"`
     Name       string `json:"name"`
     Permission string `json:"permission"`
 }
